Fix indentation of multiline strings (#9)
Thank you for this great extension.

I had one quirk that when I tried to add a multiline string, the ending
and string indentation was incorrect.

```elixir
# Current
@doc """
  My doc
  """

# Expected
@doc """
My doc
"""
```

This change seems to be fixing that and also improves indentation query
to be more readable IMO (which I did not write and all credits are [for
Helix](https://github.com/helix-editor/helix/blob/master/runtime/queries/elixir/indents.scm))