Do not delete other language servers on update (#28)
Closes https://github.com/zed-industries/zed/issues/37151

The issue here was that we were using the pattern from languages where
just one language server is available. On update, we'd remove any folder
that did not correspond to the language server currently being updated,
including these that actually belonged to other language servers.

This PR fixes this by checking folders for the correct prefix first
before deleting these.